Maintaining compatibility with existing Windows build/makefiles
