+++ lib_jtag_core/src/CMakeLists.txt
@@ -3,6 +3,7 @@ set(JTAG_BSDL_SRC
     bsdl_parser/jtag_bsdl.c
     bsdl_parser/bsdl_strings.c
     natsort/strnatcmp.c
+    dbg_logs.c
 )
 
 add_library(jtag_bsdl ${JTAG_BSDL_SRC})
